ためしにいろいろ分離してみた
